[PAXWICKET-307] Fixed serialization warning problem
The big problem was that the serialisizable checker used directly from
pax-wicket does not take into account that bundle-context and bundle are
serializable (with our own method, but still). This leads to REALLY
annoying error messages.

The additional problem was that we're not able to modifiy the wicket
SerializableChecker by extension: (a) the class was final and (b) there
where no extension points at the right places.

This patch adds a copy of the wicket SerializiableChecker, adds
additional extension points and:

(a) removes the final modfier
(b) adds a validateAdditionalSerializableConditions method which can be
overwritten (optionally) and extends the validation if an object is
serializable
(c) adds an additionalObjectReplacements method which is called during
the replaceObject stream method and allows to hook this process too.

With this patched class we're finally able to modfiy the the
SerializableChecker by errrors in the OutputStream to print the real,
internal errors. This should make working with PW ways easier.

Signed-off-by: Andreas Pieber <anpieber@gmail.com>