[PAXWICKET-350] Reverting changes which would have broken the interfaces
PAXWICKET-350 is basically dedicated to throw a servlet exception
from the filter factory BUT this would break the PAX-WICKET API which
I'm not so happy about. But those changes also include various minor
improvements to the code which are a) valuable and b) should help with
further merges to the classes. Therefore this backport had been done to
the 1.x.y branch

Signed-off-by: Andreas Pieber <anpieber@gmail.com>